updating DiCE file structure
