+++ dice_ml/explainer_interfaces/explainer_base.py
@@ -2,7 +2,7 @@
    Subclasses implement interfaces for different ML frameworks such as TensorFlow or PyTorch.
    All methods are in dice_ml.explainer_interfaces"""
 
-class DiceBase:
+class ExplainerBase:
 
     def __init__(self, data_interface):
         """Init method
